+++ apps/cli/src/commands/generate/video.ts
@@ -1,0 +1,61 @@
+import type { Command } from 'commander';
+import pc from 'picocolors';
+
+import { getTrpcClient } from '../../api/client';
+
+export function registerVideoCommand(parent: Command) {
+  parent
+    .command('video <prompt>')
+    .description('Generate a video from text')
+    .requiredOption('-m, --model <model>', 'Model ID')
+    .requiredOption('-p, --provider <provider>', 'Provider name')
+    .option('--aspect-ratio <ratio>', 'Aspect ratio (e.g. 16:9)')
+    .option('--duration <sec>', 'Duration in seconds')
+    .option('--resolution <res>', 'Resolution (e.g. 720p, 1080p)')
+    .option('--seed <n>', 'Random seed')
+    .option('--json', 'Output raw JSON')
+    .action(
+      async (
+        prompt: string,
+        options: {
+          aspectRatio?: string;
+          duration?: string;
+          json?: boolean;
+          model: string;
+          provider: string;
+          resolution?: string;
+          seed?: string;
+        },
+      ) => {
+        const client = await getTrpcClient();
+        const topicId = await client.generationTopic.createTopic.mutate({ type: 'video' });
+
+        const params: Record<string, any> = { prompt };
+        if (options.aspectRatio) params.aspectRatio = options.aspectRatio;
+        if (options.duration) params.duration = Number.parseInt(options.duration, 10);
+        if (options.resolution) params.resolution = options.resolution;
+        if (options.seed) params.seed = Number.parseInt(options.seed, 10);
+
+        const result = await client.video.createVideo.mutate({
+          generationTopicId: topicId as string,
+          model: options.model,
+          params,
+          provider: options.provider,
+        });
+
+        const r = result as any;
+        if (options.json) {
+          console.log(JSON.stringify(r, null, 2));
+          return;
+        }
+
+        console.log(`${pc.green('✓')} Video generation started`);
+        if (r.generationId) {
+          console.log(`  Generation ID: ${pc.bold(r.generationId)}`);
+        }
+        console.log(
+          pc.dim('Video generation runs asynchronously. Check status or wait for notification.'),
+        );
+      },
+    );
+}
